Refactor validator to resolve import cycles and add rule-based validation
