+++ validator/rules/fragments_on_composite_types.go
@@ -6,7 +6,7 @@ import (
 	"github.com/vektah/gqlparser/v2/ast"
 
 	//nolint:staticcheck // Validator rules each use dot imports for convenience.
-	. "github.com/vektah/gqlparser/v2/validator"
+	. "github.com/vektah/gqlparser/v2/validator/core"
 )
 
 var FragmentsOnCompositeTypesRule = Rule{
@@ -41,6 +41,3 @@ var FragmentsOnCompositeTypesRule = Rule{
 	},
 }
 
-func init() {
-	AddRule(FragmentsOnCompositeTypesRule.Name, FragmentsOnCompositeTypesRule.RuleFunc)
-}
