+++ validator/rules/known_argument_names.go
@@ -4,7 +4,7 @@ import (
 	"github.com/vektah/gqlparser/v2/ast"
 
 	//nolint:staticcheck // Validator rules each use dot imports for convenience.
-	. "github.com/vektah/gqlparser/v2/validator"
+	. "github.com/vektah/gqlparser/v2/validator/core"
 )
 
 func ruleFuncKnownArgumentNames(observers *Events, addError AddErrFunc, disableSuggestion bool) {
@@ -83,6 +83,3 @@ var KnownArgumentNamesRuleWithoutSuggestions = Rule{
 	},
 }
 
-func init() {
-	AddRule(KnownArgumentNamesRule.Name, KnownArgumentNamesRule.RuleFunc)
-}
