+++ validator/rules/known_directives.go
@@ -4,7 +4,7 @@ import (
 	"github.com/vektah/gqlparser/v2/ast"
 
 	//nolint:staticcheck // Validator rules each use dot imports for convenience.
-	. "github.com/vektah/gqlparser/v2/validator"
+	. "github.com/vektah/gqlparser/v2/validator/core"
 )
 
 var KnownDirectivesRule = Rule{
@@ -49,6 +49,3 @@ var KnownDirectivesRule = Rule{
 	},
 }
 
-func init() {
-	AddRule(KnownDirectivesRule.Name, KnownDirectivesRule.RuleFunc)
-}
