+++ validator/rules/known_fragment_names.go
@@ -4,7 +4,7 @@ import (
 	"github.com/vektah/gqlparser/v2/ast"
 
 	//nolint:staticcheck // Validator rules each use dot imports for convenience.
-	. "github.com/vektah/gqlparser/v2/validator"
+	. "github.com/vektah/gqlparser/v2/validator/core"
 )
 
 var KnownFragmentNamesRule = Rule{
@@ -21,6 +21,3 @@ var KnownFragmentNamesRule = Rule{
 	},
 }
 
-func init() {
-	AddRule(KnownFragmentNamesRule.Name, KnownFragmentNamesRule.RuleFunc)
-}
