+++ validator/rules/known_root_type.go
@@ -6,7 +6,7 @@ import (
 	"github.com/vektah/gqlparser/v2/ast"
 
 	//nolint:staticcheck // Validator rules each use dot imports for convenience.
-	. "github.com/vektah/gqlparser/v2/validator"
+	. "github.com/vektah/gqlparser/v2/validator/core"
 )
 
 var KnownRootTypeRule = Rule{
@@ -37,6 +37,3 @@ var KnownRootTypeRule = Rule{
 	},
 }
 
-func init() {
-	AddRule(KnownRootTypeRule.Name, KnownRootTypeRule.RuleFunc)
-}
