+++ validator/rules/lone_anonymous_operation.go
@@ -4,7 +4,7 @@ import (
 	"github.com/vektah/gqlparser/v2/ast"
 
 	//nolint:staticcheck // Validator rules each use dot imports for convenience.
-	. "github.com/vektah/gqlparser/v2/validator"
+	. "github.com/vektah/gqlparser/v2/validator/core"
 )
 
 var LoneAnonymousOperationRule = Rule{
@@ -21,6 +21,3 @@ var LoneAnonymousOperationRule = Rule{
 	},
 }
 
-func init() {
-	AddRule(LoneAnonymousOperationRule.Name, LoneAnonymousOperationRule.RuleFunc)
-}
