+++ validator/rules/max_introspection_depth.go
@@ -4,7 +4,7 @@ import (
 	"github.com/vektah/gqlparser/v2/ast"
 
 	//nolint:staticcheck // Validator rules each use dot imports for convenience.
-	. "github.com/vektah/gqlparser/v2/validator"
+	. "github.com/vektah/gqlparser/v2/validator/core"
 )
 
 const maxListsDepth = 3
@@ -85,6 +85,3 @@ func checkDepthFragmentSpread(fragmentSpread *ast.FragmentSpread, visitedFragmen
 	return checkDepthSelectionSet(fragment.SelectionSet, visitedFragments, depth)
 }
 
-func init() {
-	AddRule(MaxIntrospectionDepth.Name, MaxIntrospectionDepth.RuleFunc)
-}
