+++ validator/rules/no_fragment_cycles.go
@@ -7,7 +7,7 @@ import (
 	"github.com/vektah/gqlparser/v2/ast"
 
 	//nolint:staticcheck // Validator rules each use dot imports for convenience.
-	. "github.com/vektah/gqlparser/v2/validator"
+	. "github.com/vektah/gqlparser/v2/validator/core"
 )
 
 var NoFragmentCyclesRule = Rule{
@@ -71,9 +71,6 @@ var NoFragmentCyclesRule = Rule{
 	},
 }
 
-func init() {
-	AddRule(NoFragmentCyclesRule.Name, NoFragmentCyclesRule.RuleFunc)
-}
 
 func getFragmentSpreads(node ast.SelectionSet) []*ast.FragmentSpread {
 	var spreads []*ast.FragmentSpread
