+++ validator/rules/no_undefined_variables.go
@@ -4,7 +4,7 @@ import (
 	"github.com/vektah/gqlparser/v2/ast"
 
 	//nolint:staticcheck // Validator rules each use dot imports for convenience.
-	. "github.com/vektah/gqlparser/v2/validator"
+	. "github.com/vektah/gqlparser/v2/validator/core"
 )
 
 var NoUndefinedVariablesRule = Rule{
@@ -30,6 +30,3 @@ var NoUndefinedVariablesRule = Rule{
 	},
 }
 
-func init() {
-	AddRule(NoUndefinedVariablesRule.Name, NoUndefinedVariablesRule.RuleFunc)
-}
