+++ validator/rules/no_unused_fragments.go
@@ -4,7 +4,7 @@ import (
 	"github.com/vektah/gqlparser/v2/ast"
 
 	//nolint:staticcheck // Validator rules each use dot imports for convenience.
-	. "github.com/vektah/gqlparser/v2/validator"
+	. "github.com/vektah/gqlparser/v2/validator/core"
 )
 
 var NoUnusedFragmentsRule = Rule{
@@ -31,6 +31,3 @@ var NoUnusedFragmentsRule = Rule{
 	},
 }
 
-func init() {
-	AddRule(NoUnusedFragmentsRule.Name, NoUnusedFragmentsRule.RuleFunc)
-}
