+++ validator/rules/no_unused_variables.go
@@ -4,7 +4,7 @@ import (
 	"github.com/vektah/gqlparser/v2/ast"
 
 	//nolint:staticcheck // Validator rules each use dot imports for convenience.
-	. "github.com/vektah/gqlparser/v2/validator"
+	. "github.com/vektah/gqlparser/v2/validator/core"
 )
 
 var NoUnusedVariablesRule = Rule{
@@ -32,6 +32,3 @@ var NoUnusedVariablesRule = Rule{
 	},
 }
 
-func init() {
-	AddRule(NoUnusedVariablesRule.Name, NoUnusedVariablesRule.RuleFunc)
-}
