+++ validator/rules/overlapping_fields_can_be_merged.go
@@ -8,7 +8,7 @@ import (
 	"github.com/vektah/gqlparser/v2/ast"
 
 	//nolint:staticcheck // Validator rules each use dot imports for convenience.
-	. "github.com/vektah/gqlparser/v2/validator"
+	. "github.com/vektah/gqlparser/v2/validator/core"
 )
 
 var OverlappingFieldsCanBeMergedRule = Rule{
@@ -108,9 +108,6 @@ var OverlappingFieldsCanBeMergedRule = Rule{
 	},
 }
 
-func init() {
-	AddRule(OverlappingFieldsCanBeMergedRule.Name, OverlappingFieldsCanBeMergedRule.RuleFunc)
-}
 
 type pairSet struct {
 	data map[string]map[string]bool
