+++ validator/rules/possible_fragment_spreads.go
@@ -4,7 +4,7 @@ import (
 	"github.com/vektah/gqlparser/v2/ast"
 
 	//nolint:staticcheck // Validator rules each use dot imports for convenience.
-	. "github.com/vektah/gqlparser/v2/validator"
+	. "github.com/vektah/gqlparser/v2/validator/core"
 )
 
 var PossibleFragmentSpreadsRule = Rule{
@@ -69,6 +69,3 @@ var PossibleFragmentSpreadsRule = Rule{
 	},
 }
 
-func init() {
-	AddRule(PossibleFragmentSpreadsRule.Name, PossibleFragmentSpreadsRule.RuleFunc)
-}
