+++ validator/rules/provided_required_arguments.go
@@ -3,7 +3,7 @@ package rules
 import (
 	"github.com/vektah/gqlparser/v2/ast"
 	//nolint:staticcheck // Validator rules each use dot imports for convenience.
-	. "github.com/vektah/gqlparser/v2/validator"
+	. "github.com/vektah/gqlparser/v2/validator/core"
 )
 
 var ProvidedRequiredArgumentsRule = Rule{
@@ -63,6 +63,3 @@ var ProvidedRequiredArgumentsRule = Rule{
 	},
 }
 
-func init() {
-	AddRule(ProvidedRequiredArgumentsRule.Name, ProvidedRequiredArgumentsRule.RuleFunc)
-}
