+++ validator/rules/single_field_subscriptions.go
@@ -7,7 +7,7 @@ import (
 	"github.com/vektah/gqlparser/v2/ast"
 
 	//nolint:staticcheck // Validator rules each use dot imports for convenience.
-	. "github.com/vektah/gqlparser/v2/validator"
+	. "github.com/vektah/gqlparser/v2/validator/core"
 )
 
 var SingleFieldSubscriptionsRule = Rule{
@@ -44,9 +44,6 @@ var SingleFieldSubscriptionsRule = Rule{
 	},
 }
 
-func init() {
-	AddRule(SingleFieldSubscriptionsRule.Name, SingleFieldSubscriptionsRule.RuleFunc)
-}
 
 type topField struct {
 	name     string
