+++ validator/rules/unique_argument_names.go
@@ -4,7 +4,7 @@ import (
 	"github.com/vektah/gqlparser/v2/ast"
 
 	//nolint:staticcheck // Validator rules each use dot imports for convenience.
-	. "github.com/vektah/gqlparser/v2/validator"
+	. "github.com/vektah/gqlparser/v2/validator/core"
 )
 
 var UniqueArgumentNamesRule = Rule{
@@ -20,9 +20,6 @@ var UniqueArgumentNamesRule = Rule{
 	},
 }
 
-func init() {
-	AddRule(UniqueArgumentNamesRule.Name, UniqueArgumentNamesRule.RuleFunc)
-}
 
 func checkUniqueArgs(args ast.ArgumentList, addError AddErrFunc) {
 	knownArgNames := map[string]int{}
