+++ validator/rules/unique_directives_per_location.go
@@ -4,7 +4,7 @@ import (
 	"github.com/vektah/gqlparser/v2/ast"
 
 	//nolint:staticcheck // Validator rules each use dot imports for convenience.
-	. "github.com/vektah/gqlparser/v2/validator"
+	. "github.com/vektah/gqlparser/v2/validator/core"
 )
 
 var UniqueDirectivesPerLocationRule = Rule{
@@ -26,6 +26,3 @@ var UniqueDirectivesPerLocationRule = Rule{
 	},
 }
 
-func init() {
-	AddRule(UniqueDirectivesPerLocationRule.Name, UniqueDirectivesPerLocationRule.RuleFunc)
-}
