+++ validator/rules/unique_fragment_names.go
@@ -4,7 +4,7 @@ import (
 	"github.com/vektah/gqlparser/v2/ast"
 
 	//nolint:staticcheck // Validator rules each use dot imports for convenience.
-	. "github.com/vektah/gqlparser/v2/validator"
+	. "github.com/vektah/gqlparser/v2/validator/core"
 )
 
 var UniqueFragmentNamesRule = Rule{
@@ -24,6 +24,3 @@ var UniqueFragmentNamesRule = Rule{
 	},
 }
 
-func init() {
-	AddRule(UniqueFragmentNamesRule.Name, UniqueFragmentNamesRule.RuleFunc)
-}
