+++ validator/rules/unique_input_field_names.go
@@ -4,7 +4,7 @@ import (
 	"github.com/vektah/gqlparser/v2/ast"
 
 	//nolint:staticcheck // Validator rules each use dot imports for convenience.
-	. "github.com/vektah/gqlparser/v2/validator"
+	. "github.com/vektah/gqlparser/v2/validator/core"
 )
 
 var UniqueInputFieldNamesRule = Rule{
@@ -29,6 +29,3 @@ var UniqueInputFieldNamesRule = Rule{
 	},
 }
 
-func init() {
-	AddRule(UniqueInputFieldNamesRule.Name, UniqueInputFieldNamesRule.RuleFunc)
-}
