+++ validator/rules/variables_are_input_types.go
@@ -4,7 +4,7 @@ import (
 	"github.com/vektah/gqlparser/v2/ast"
 
 	//nolint:staticcheck // Validator rules each use dot imports for convenience.
-	. "github.com/vektah/gqlparser/v2/validator"
+	. "github.com/vektah/gqlparser/v2/validator/core"
 )
 
 var VariablesAreInputTypesRule = Rule{
@@ -30,6 +30,3 @@ var VariablesAreInputTypesRule = Rule{
 	},
 }
 
-func init() {
-	AddRule(VariablesAreInputTypesRule.Name, VariablesAreInputTypesRule.RuleFunc)
-}
