+++ validator/rules/variables_in_allowed_position.go
@@ -4,7 +4,7 @@ import (
 	"github.com/vektah/gqlparser/v2/ast"
 
 	//nolint:staticcheck // Validator rules each use dot imports for convenience.
-	. "github.com/vektah/gqlparser/v2/validator"
+	. "github.com/vektah/gqlparser/v2/validator/core"
 )
 
 var VariablesInAllowedPositionRule = Rule{
@@ -40,6 +40,3 @@ var VariablesInAllowedPositionRule = Rule{
 	},
 }
 
-func init() {
-	AddRule(VariablesInAllowedPositionRule.Name, VariablesInAllowedPositionRule.RuleFunc)
-}
